# [SPARK-31606][SQL] Reduce the perf regression of vectorized parquet reader caused by datetime rebase