--- a/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/DateTimeRebaseBenchmark.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/DateTimeRebaseBenchmark.scala
@@ -49,15 +49,15 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
       .select($"seconds".cast("timestamp").as("ts"))
   }
 
-  private def genTsAfter1582(cardinality: Int): DataFrame = {
-    val start = LocalDateTime.of(1582, 10, 15, 0, 0, 0)
+  private def genTsAfter1900(cardinality: Int): DataFrame = {
+    val start = LocalDateTime.of(1900, 1, 31, 0, 0, 0)
     val end = LocalDateTime.of(3000, 1, 1, 0, 0, 0)
     genTs(cardinality, start, end)
   }
 
-  private def genTsBefore1582(cardinality: Int): DataFrame = {
+  private def genTsBefore1900(cardinality: Int): DataFrame = {
     val start = LocalDateTime.of(10, 1, 1, 0, 0, 0)
-    val end = LocalDateTime.of(1580, 1, 1, 0, 0, 0)
+    val end = LocalDateTime.of(1900, 1, 1, 0, 0, 0)
     genTs(cardinality, start, end)
   }
 
@@ -71,34 +71,35 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
   }
 
   private def genDateAfter1582(cardinality: Int): DataFrame = {
-    val start = LocalDate.of(1582, 10, 15)
+    val start = LocalDate.of(1582, 10, 31)
     val end = LocalDate.of(3000, 1, 1)
     genDate(cardinality, start, end)
   }
 
   private def genDateBefore1582(cardinality: Int): DataFrame = {
     val start = LocalDate.of(10, 1, 1)
-    val end = LocalDate.of(1580, 1, 1)
+    val end = LocalDate.of(1580, 10, 1)
     genDate(cardinality, start, end)
   }
 
-  private def genDF(cardinality: Int, dateTime: String, after1582: Boolean): DataFrame = {
-    (dateTime, after1582) match {
+  private def genDF(cardinality: Int, dateTime: String, modernDates: Boolean): DataFrame = {
+    (dateTime, modernDates) match {
       case ("date", true) => genDateAfter1582(cardinality)
       case ("date", false) => genDateBefore1582(cardinality)
-      case ("timestamp", true) => genTsAfter1582(cardinality)
-      case ("timestamp", false) => genTsBefore1582(cardinality)
+      case ("timestamp", true) => genTsAfter1900(cardinality)
+      case ("timestamp", false) => genTsBefore1900(cardinality)
       case _ => throw new IllegalArgumentException(
-        s"cardinality = $cardinality dateTime = $dateTime after1582 = $after1582")
+        s"cardinality = $cardinality dateTime = $dateTime modernDates = $modernDates")
     }
   }
 
   private def benchmarkInputs(benchmark: Benchmark, rowsNum: Int, dateTime: String): Unit = {
-    benchmark.addCase("after 1582, noop", 1) { _ =>
-      genDF(rowsNum, dateTime, after1582 = true).noop()
+    val year = if (dateTime == "date") 1582 else 1900
+    benchmark.addCase(s"after $year, noop", 1) { _ =>
+      genDF(rowsNum, dateTime, modernDates = true).noop()
     }
-    benchmark.addCase("before 1582, noop", 1) { _ =>
-      genDF(rowsNum, dateTime, after1582 = false).noop()
+    benchmark.addCase(s"before $year, noop", 1) { _ =>
+      genDF(rowsNum, dateTime, modernDates = false).noop()
     }
   }
 
@@ -107,23 +108,26 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
   }
 
   private def caseName(
-      after1582: Boolean,
+      modernDates: Boolean,
+      dateTime: String,
       rebase: Option[Boolean] = None,
       vec: Option[Boolean] = None): String = {
-    val period = if (after1582) "after" else "before"
+    val period = if (modernDates) "after" else "before"
+    val year = if (dateTime == "date") 1582 else 1900
     val vecFlag = vec.map(flagToStr).map(flag => s", vec $flag").getOrElse("")
     val rebaseFlag = rebase.map(flagToStr).map(flag => s", rebase $flag").getOrElse("")
-    s"$period 1582$vecFlag$rebaseFlag"
+    s"$period $year$vecFlag$rebaseFlag"
   }
 
   private def getPath(
       basePath: File,
       dateTime: String,
-      after1582: Boolean,
+      modernDates: Boolean,
       rebase: Option[Boolean] = None): String = {
-    val period = if (after1582) "after" else "before"
+    val period = if (modernDates) "after" else "before"
+    val year = if (dateTime == "date") 1582 else 1900
     val rebaseFlag = rebase.map(flagToStr).map(flag => s"_$flag").getOrElse("")
-    basePath.getAbsolutePath + s"/${dateTime}_${period}_1582$rebaseFlag"
+    basePath.getAbsolutePath + s"/${dateTime}_${period}_$year$rebaseFlag"
   }
 
   override def runBenchmarkSuite(mainArgs: Array[String]): Unit = {
@@ -139,16 +143,16 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
                 rowsNum,
                 output = output)
               benchmarkInputs(benchmark, rowsNum, dateTime)
-              Seq(true, false).foreach { after1582 =>
+              Seq(true, false).foreach { modernDates =>
                 Seq(false, true).foreach { rebase =>
-                  benchmark.addCase(caseName(after1582, Some(rebase)), 1) { _ =>
+                  benchmark.addCase(caseName(modernDates, dateTime, Some(rebase)), 1) { _ =>
                     withSQLConf(
                       SQLConf.LEGACY_PARQUET_REBASE_DATETIME_IN_WRITE.key -> rebase.toString) {
-                      genDF(rowsNum, dateTime, after1582)
+                      genDF(rowsNum, dateTime, modernDates)
                         .write
                         .mode("overwrite")
                         .format("parquet")
-                        .save(getPath(path, dateTime, after1582, Some(rebase)))
+                        .save(getPath(path, dateTime, modernDates, Some(rebase)))
                     }
                   }
                 }
@@ -157,16 +161,15 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
 
               val benchmark2 = new Benchmark(
                 s"Load ${dateTime}s from parquet", rowsNum, output = output)
-              Seq(true, false).foreach { after1582 =>
+              Seq(true, false).foreach { modernDates =>
                 Seq(false, true).foreach { vec =>
                   Seq(false, true).foreach { rebase =>
-                    benchmark2.addCase(caseName(after1582, Some(rebase), Some(vec)), 3) { _ =>
-                      withSQLConf(
-                        SQLConf.PARQUET_VECTORIZED_READER_ENABLED.key -> vec.toString,
-                        SQLConf.LEGACY_PARQUET_REBASE_DATETIME_IN_READ.key -> rebase.toString) {
+                    val name = caseName(modernDates, dateTime, Some(rebase), Some(vec))
+                    benchmark2.addCase(name, 3) { _ =>
+                      withSQLConf(SQLConf.PARQUET_VECTORIZED_READER_ENABLED.key -> vec.toString) {
                         spark.read
                           .format("parquet")
-                          .load(getPath(path, dateTime, after1582, Some(rebase)))
+                          .load(getPath(path, dateTime, modernDates, Some(rebase)))
                           .noop()
                       }
                     }
@@ -183,13 +186,13 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
             Seq("date", "timestamp").foreach { dateTime =>
               val benchmark = new Benchmark(s"Save ${dateTime}s to ORC", rowsNum, output = output)
               benchmarkInputs(benchmark, rowsNum, dateTime)
-              Seq(true, false).foreach { after1582 =>
-                benchmark.addCase(caseName(after1582), 1) { _ =>
-                  genDF(rowsNum, dateTime, after1582)
+              Seq(true, false).foreach { modernDates =>
+                benchmark.addCase(caseName(modernDates, dateTime), 1) { _ =>
+                  genDF(rowsNum, dateTime, modernDates)
                     .write
                     .mode("overwrite")
                     .format("orc")
-                    .save(getPath(path, dateTime, after1582))
+                    .save(getPath(path, dateTime, modernDates))
                 }
               }
               benchmark.run()
@@ -198,14 +201,14 @@ object DateTimeRebaseBenchmark extends SqlBasedBenchmark {
                 s"Load ${dateTime}s from ORC",
                 rowsNum,
                 output = output)
-              Seq(true, false).foreach { after1582 =>
+              Seq(true, false).foreach { modernDates =>
                 Seq(false, true).foreach { vec =>
-                  benchmark2.addCase(caseName(after1582, vec = Some(vec)), 3) { _ =>
+                  benchmark2.addCase(caseName(modernDates, dateTime, vec = Some(vec)), 3) { _ =>
                     withSQLConf(SQLConf.ORC_VECTORIZED_READER_ENABLED.key -> vec.toString) {
                       spark
                         .read
                         .format("orc")
-                        .load(getPath(path, dateTime, after1582))
+                        .load(getPath(path, dateTime, modernDates))
                         .noop()
                     }
                   }