@MaxGekk 2020-04-29
@cloud-fan Please, review and merge https://github.com/cloud-fan/spark/pull/15