@HyukjinKwon 2020-05-04
Merged to master and branch-3.0.